pyproject.toml
[project]
name = "agent-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.12.2",
    "boto3==1.34.11",
    "chromadb>=1.3.5",
    "cryptography==41.0.7",
    "faiss-cpu==1.7.4",
    "fastapi==0.108.0",
    "finance-datareader>=0.9.96",
    "httpx==0.27.2",
    "langchain==0.2.0",
    "langchain-core>=0.2.12",
    "langchain-openai==0.1.12",
    "langgraph==0.2.3",
    "langgraph-checkpoint==1.0.2",
    "langgraph-checkpoint-postgres==1.0.3",
    "langgraph-checkpoint-sqlite==1.0.0",
    "newspaper3k==0.2.8",
    "numpy==1.26.2",
    "openai==1.40.0",
    "pandas==2.1.4",
    "pydantic>=2.0.0",
    "pypdf==3.17.4",
    "python-dotenv>=1.0.0",
    "qdrant-client==1.7.0",
    "redis==5.0.1",
    "requests==2.31.0",
    "scipy==1.11.4",
    "ta==0.11.0",
    "uvicorn==0.25.0",
    "yfinance==0.2.33",
]
